src/api/routes/documents/documents-ckeditor.php src/EcclesiaCRM/APIControllers/DocumentCKEditorController.php
